perf(home): memoize feed rows, bound group expansion; extract row layer
HomeFeedRow/EndorsementGroupRow/HomeFeedBody memoized (element-wise group
comparator per review); grouped-endorsement expansion windows 50 accounts
per click instead of mounting up to ~1000; row/sentence/preview layer moved
verbatim to home-feed-rows.tsx (CertPreview re-exported for tests).
